% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shortenLabel}
\alias{shortenLabel}
\title{Shorten labels for display}
\usage{
shortenLabel(x, n = 16)
}
\arguments{
\item{x}{Character vector of labels.}

\item{n}{Maximum number of characters.}
}
\description{
\code{shortenLabel} truncates long labels with an ellipsis so they fit
inside path diagram boxes.
}
\keyword{internal}
