% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sigStars}
\alias{sigStars}
\title{Significance stars for p values}
\usage{
sigStars(p)
}
\arguments{
\item{p}{Numeric vector of p values.}
}
\description{
\code{sigStars} converts p values to the conventional significance stars:
\code{*} p < .05, \code{**} p < .01, \code{***} p < .001.
}
