% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medmod.R
\docType{package}
\name{medmod-package}
\alias{medmod}
\alias{medmod-package}
\title{medmod: Simple Mediation and Moderation Analysis}
\description{
A package for simple mediation and moderation based on the \code{\link[lavaan]{lavaan}}
package by Yves Rosseel. Medmod is also available as a
module for 'jamovi' (see \href{https://www.jamovi.org}{www.jamovi.org} for
more information).
}
\details{
\tabular{llllll}{
  Simple mediation analysis \tab \tab \tab \tab \code{\link{med}()} \cr
  Simple moderation analysis \tab \tab \tab \tab \code{\link{mod}()} \cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/raviselker/medmod}
  \item Report bugs at \url{https://github.com/raviselker/medmod/issues}
}

}
