% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedMod.R
\name{MedMod}
\alias{MedMod}
\title{Mediated moderation analysis with the debiased machine learning method}
\usage{
MedMod(
  data,
  outcome,
  mediators,
  treatment,
  subgroup,
  covariates,
  learners = c("SL.glm"),
  learners_h = NULL,
  learners_mu = NULL,
  num_folds = 5,
  ci.level = 0.95
)
}
\arguments{
\item{data}{a \code{data.frame} containing variables involved in the analysis.}

\item{outcome}{a character string of the name of the column in "data" that correspond to the outcome variable (Y).}

\item{mediators}{a character vector of the names of the columns in "data" that correspond to the mediator M1. The definition for the mediated moderation via mediator M1 (i.e., MedMod_M1) considers shifting the levels of M1, while fixing the levels of mediator M2 at its levels among the reference subgroup.}

\item{treatment}{a character string of the name of the column in "data" that correspond to the treatment variable (T). The treatment variable should be dummy coded (e.g., 1 for the intervention condition, 0 for the control condition).}

\item{subgroup}{a character string of the name of the column in "data" that correspond to the moderator subgroup status (R). The subgroup status should be dummy coded, with 1 for the focal subgroup and 0 for the reference subgroup.}

\item{covariates}{a character vector of the names of the columns in "data" that correspond to pretreatment (i.e., baseline) covariates (C).}

\item{learners}{a character vector specifying the methods for estimating the nuisance models. Currently supported options include the list of methods included in the "SuperLearner" package (<https://cran.r-project.org/package=SuperLearner>), which can be listed via function listWrappers() of the "SuperLearner" package.}

\item{num_folds}{the number of folds used for the cross-fitting procedure.}
}
\value{
A data frame containing the results for the mediated moderation estimands and the expected potential outcomes in the estimands. The results include the estimates, standard error estimates, and 0.95 confidence intervals.
}
\description{
Mediated moderation analysis with the debiased machine learning method
}
\examples{

library(tidyverse)
library(glue)
library(origami)
library(mvtnorm)
library(SuperLearner)
library(ranger)
library(nnet)

# Run ----------
data <- MedMod::data
head(data)

set.seed(12345)
out <- MedMod::MedMod(
  data = data, # data containing all variables
  outcome = "Outcome", # name of outcome
  mediators = c("M1", "M2"), # name of mediators. When more than two mediators are included, the first mediator is designated as M1, and all remaining mediators are collectively treated as M2.
  treatment = "Intervention", # name of treatment variable
  subgroup = "Male", # name of moderator subgroup variable
  covariates = c("C.1", "C.2", "C.3"), # names of baseline covariates
  learners = c("SL.mean", "SL.glm", "SL.ranger", "SL.nnet"),
  # methods from the SuperLearner package to include in the super learner ensemble (intercept-only model, generalized linear model, random forest, neural network).
  # To see all available methods, run: SuperLearner::listWrappers()
  num_folds = 4, # number of folds for cross-fitting
  ci.level = 0.95 # default: 95\% confidence interval
)
# Extract results for mediated moderation
out \%>\%
  filter(Estimand \%in\% c("TotMod", "MedMod", "RemainMod", # Total moderation, Mediated moderation, Remaining  moderation
                         "MedMod_M1", "MedMod_M2", "MedMod_mu")) # When more than one mediators, the output also include: Mediated moderation via M1, Mediated moderation via M2, Mediated moderation due to mediators' mutual dependence


}
